+++ pytsite/auth_ui/res/assets/js/widget/follow.js
@@ -28,7 +28,7 @@ $(window).on('pytsite.widget.init:pytsite.auth_ui._widget.Follow', function (e, 
             if (!confirm(t('pytsite.auth_ui@unfollow_confirmation')))
                 return;
 
-            pytsite.jsApi.post('pytsite.auth_ui.follow', {
+            pytsite.jsApi.post('pytsite.auth_ui@follow', {
                 op: 'unfollow',
                 uid: em.data('userId')
             }).done(function (data) {
@@ -40,7 +40,7 @@ $(window).on('pytsite.widget.init:pytsite.auth_ui._widget.Follow', function (e, 
             });
         }
         else if (btn.hasClass('non-following')) {
-            pytsite.jsApi.post('pytsite.auth_ui.follow', {
+            pytsite.jsApi.post('pytsite.auth_ui@follow', {
                 op: 'follow',
                 uid: em.data('userId')
             }).done(function (data) {