+++ pytsite/flag/js_api.py
@@ -8,7 +8,7 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
-def like(args: dict, inp: dict) -> dict:
+def like(inp: dict) -> dict:
     current_user = _auth.get_current_user()
     entity = _odm.get_by_ref(inp.get('entity'))